LayoutTests/http/tests/websocket/tests/hybi/unpaired-surrogates-in-message.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Checks whether SYNTAX_ERR is thrown when attemping to send unpaired surrogates.");

window.jsTestIsAsync = true;
if (window.testRunner)
    testRunner.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var ws = new WebSocket("ws://localhost:8880/websocket/tests/hybi/echo");

ws.onopen = function()
{
    debug("Connected.");
    try {
        ws.send('\uD807');
        testFailed('SYNTAX_ERR should be thrown.');
    } catch(e) {
        if (e.name == 'SYNTAX_ERR')
            testPassed('SYNTAX_ERR was thrown.');
        else
            testFailed('Unexpected exception: ' + e);
    }
    ws.close();
};

ws.onmessage = function (event)
{
    var message = event.data;
    testFailed("onmessage() was called. (message = \"" + message + "\")");
    ws.close();
};

ws.onclose = function()
{
    finishJSTest();
};

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
